% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_calculate_ATS.R
\name{calculate_ATS}
\alias{calculate_ATS}
\title{Calculate ATS}
\usage{
calculate_ATS(
  chartij,
  ttij,
  nobs,
  starttime,
  endtime,
  ttmin,
  ttmax,
  ntimepoints,
  control_limit,
  no_signal_action = "omit"
)
}
\arguments{
\item{chartij}{a matrix of charting statistics. 
\code{chartij[i,j]} is the jth charting statistic of the ith subject.}

\item{ttij}{a matrix of observation times.
\code{ttij[i,j]} is the jth observation time of the ith subject.}

\item{nobs}{an integer vector for number of observations.
\code{nobs[i]} is the number of observations for the ith subject.}

\item{starttime}{a vector of times from the start of monitoring.
\code{starttime[i]} is the time that the ith subject starts to be monitored.}

\item{endtime}{a vector of times from the start of monitoring.
\code{endtime[i]} is the time that the ith subject is lost to be monitored.}

\item{ttmin, ttmax}{two numeric values.
[\code{ttmin},\code{ttmax}] is the design time interval.}

\item{ntimepoints}{an integer value.
\code{ntimepoints} is the number of basic timepoints in the design time interval.}

\item{control_limit}{a numeric value.
\code{control_limit} is the control limit, signals will be given if charting statistics are greater than the control limit.}
}
\value{
ATS, a numeric value.
}
\description{
Calculate ATS
}
\author{
LY
}
