% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cleaning.R
\name{clean_matij_by_nobs}
\alias{clean_matij_by_nobs}
\title{Ancillary Function for Cleaning Matrices}
\usage{
clean_matij_by_nobs(matij, nobs, object_name = "matij")
}
\arguments{
\item{matij}{a matrix to be cleaned by \code{nobs}. \cr
\code{matij[i,j]} is the jth observation of the ith subject.}

\item{nobs}{an integer vector of observation times. \cr
\code{nobs[i]} is the number of observations for the ith subject.}
}
\value{
a cleaned matrix of \code{matij}. \cr
\code{matij[i,j]} is set to NA if \code{j>nobs[i]}
}
\description{
Ancillary Function for Cleaning Matrices
}
\author{
LY
}
