% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pattern_long_1d.R
\name{est_pattern_long_1d}
\alias{est_pattern_long_1d}
\title{Estimate the Pattern of Univariate Longitudinal Data}
\usage{
est_pattern_long_1d(
  yyij,
  ttij,
  nobs,
  ttmin,
  ttmax,
  ntimepoints,
  method,
  smoothing = "local linear",
  hh_mean,
  hh_var,
  hh_cov,
  hh_t,
  hh_y
)
}
\arguments{
\item{yyij}{a matrix of longitudinal observations. \cr
\code{yyij[i,j]} is the jth observation of the kth dimension of the ith subject.}

\item{ttij}{a matrix of observation times. \cr
\code{ttij[i,j]} is the jth observation time of the ith subject. \cr
\code{yyij[i,j]} is observed at \code{ttij[i,j]}.}

\item{nobs}{an integer vector for number of observations. \cr
\code{nobs[i]} is the number of observations for the ith subject.}

\item{ttmin, ttmax}{two numeric values. \cr
[\code{ttmin},\code{ttmax}] is the design time interval.}

\item{ntimepoints}{an integer value. \cr
\code{ntimepoints} is the number of basic timepoints in the design time interval.}

\item{method}{a string. \cr
If \code{method="meanvar"}, the function will estimate the \cr
If \code{method="meanvarcov"}, the function will estimate the \cr
If \code{method="meanvarcovmean"}, the function will estimate the \cr
If \code{method="distribution"}, the function will estimate the \cr
If \code{method="distributionvarcov"}, the function will estimate the}

\item{smoothing}{a string. \cr
If \code{smoothing="local constant"}, apply local constant approximation \cr
If \code{smoothing="local linear"}, apply local linear approximation}

\item{hh_mean}{an integer value. \cr
The bandwidth parameter for estimating mean function.}

\item{hh_var}{an integer value. \cr
The bandwidth parameter for estimating variance function.}

\item{hh_cov}{an integer value. \cr
The bandwidth parameter for estimating covariance function.}

\item{hh_t}{an integer value. \cr
The bandwidth parameter in time axis for estimating distribution function.}

\item{hh_y}{a numeric value. \cr
The bandwidth parameter in y-axis for estimating distribution function.}
}
\description{
Estimate the Pattern of Univariate Longitudinal Data
}
