% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pattern_long_md.R
\name{est_pattern_long_md}
\alias{est_pattern_long_md}
\title{Estimate the Pattern of Multivariate Longitudinal Data}
\usage{
est_pattern_long_md(
  yyijk,
  ttij,
  nobs,
  ttmin,
  ttmax,
  ntimepoints,
  method,
  hh_mean,
  hh_var,
  hh_cov
)
}
\arguments{
\item{yyijk}{a 3d-array of longitudinal observations. \cr
\code{yyijk[i,j,k]} is the jth observation of the kth dimension of the ith subject.}

\item{ttij}{a matrix of observation times. \cr
\code{ttij[i,j]} is the jth observation time of the ith subject. \cr
\code{yyijk[i,j,]} is observed at \code{ttij[i,j]}.}

\item{nobs}{an integer vector for number of observations. \cr
\code{nobs[i]} is the number of observations for the ith subject.}

\item{ttmin, ttmax}{two numeric values. \cr
[\code{ttmin},\code{ttmax}] is the design time interval.}

\item{ntimepoints}{an integer value. \cr
\code{ntimepoints} is the number of basic timepoints in the design time interval.}

\item{method}{a string. \cr
If \code{method="meanvar"}, the function will estimate the \cr
mean function (\eqn{\mathrm{E}[\mathbf{y}(t)]}), and \cr
variance function (\eqn{\mathrm{Var}(\mathbf{y}(t))}). \cr
If \code{method="meanvarcov"}, the function will estimate the \cr
mean function (\eqn{\mathrm{E}[\mathbf{y}(t)]}), \cr
variance function (\eqn{\mathrm{Var}(\mathbf{y}(t))}), and \cr
covariance function (\eqn{\mathrm{Cov}(\mathbf{y}(s),\mathbf{y}(t))}).}

\item{hh_mean}{an integer value. \cr
The bandwidth parameter for estimating mean function}

\item{hh_var}{an integer value. \cr
The bandwidth parameter for estimating variance function}

\item{hh_cov}{an integer value. \cr
The bandwidth parameter for estimating covariance function}
}
\value{
the estimated longitudinal pattern. \cr
If \code{method="meanvar"}, return an object of class \code{patn_long_mult_meanvar}. \cr
If \code{method="meanvarcov"}, return an object of class \code{patn_long_mult_meanvarcov}.
}
\description{
Estimate the Pattern of Multivariate Longitudinal Data
}
\author{
LY
}
