% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pattern_long_surv.R
\name{est_pattern_long_surv}
\alias{est_pattern_long_surv}
\title{Estimate the Pattern of Longitudinal and Survival Data}
\usage{
est_pattern_long_surv(
  yyijk,
  ttij,
  nobs,
  survtime,
  survevent,
  ttmin,
  ttmax,
  ntimepoints,
  method = "risk",
  smoothing = "local linear",
  hh_beta,
  hh_mean,
  hh_var,
  hh
)
}
\arguments{
\item{yyijk}{a 3d-array of longitudinal observations. \cr
\code{yyijk[i,j,k]} is the jth observation of the kth dimension of the ith subject.}

\item{ttij}{a matrix of observation times. \cr
\code{ttij[i,j]} is the jth observation time of the ith subject. \cr
\code{yyijk[i,j,]} is observed at \code{ttij[i,j]}.}

\item{nobs}{an integer vector for number of observations. \cr
\code{nobs[i]} is the number of observations for the ith subject.}

\item{survtime}{a vector of survival times \cr
\code{survtime[i]} is the survival time of the ith subject.}

\item{survevent}{a logical vector of survival events \cr
If \code{survevents[i]==TRUE}, then a survival event is observed at \code{survtime[i]}. \cr
If \code{survevents[i]==FALSE}, then no survival event is observed at \code{survtime[i]}.}

\item{ttmin, ttmax}{two numeric values. \cr
[\code{ttmin},\code{ttmax}] is the design time interval.}

\item{ntimepoints}{an integer value. \cr
\code{ntimepoints} is the number of basic timepoints in the design time interval.}

\item{method}{a string. \cr
\code{method} is the number of basic timepoints in the design time interval. \cr
If \code{method="risk"}, \cr
If \code{method="joint"},}

\item{smoothing}{a string. \cr
If \code{smoothing="local constant"}, apply local constant approximation \cr
If \code{smoothing="local linear"}, apply local linear approximation \cr
If \code{smoothing="local quadratic"}, apply local quadratic approximation}

\item{hh_beta}{an integer value. \cr
The bandwidth parameter for estimating the regression coefficients beta in the Cox model.}

\item{hh_mean}{an integer value. \cr
The bandwidth parameter for estimating mean function.}

\item{hh_var}{an integer value. \cr
The bandwidth parameter for estimating variance function.}

\item{hh_cov}{an integer value. \cr
The bandwidth parameter for estimating covariance function.}

\item{hh_t}{an integer value. \cr
The bandwidth parameter in time axis for estimating distribution function.}

\item{hh_y}{a numeric value. \cr
The bandwidth parameter in y-axis for estimating distribution function.}
}
\description{
Estimate the Pattern of Longitudinal and Survival Data
}
