% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\docType{data}
\name{example_long_md}
\alias{example_long_md}
\title{Data Example: Multivariate Longitudinal Data}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(example_long_md)
}
\description{
A similated multivariate longitudinal data set for demonstration and testing.
}
\examples{
data(example_long_md)
}
\keyword{datasets}
