% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\docType{data}
\name{example_long_surv}
\alias{example_long_surv}
\title{Data Example: Longitudinal and Survival Data}
\format{
An object of class \code{list} of length 6.
}
\usage{
data(example_long_surv)
}
\description{
A similated data set with longitudinal and survival data for demonstration and testing.
}
\examples{
data(example_long_surv)
}
\keyword{datasets}
