% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnt_long_1d.R
\name{mnt_long_1d}
\alias{mnt_long_1d}
\title{Monitor Univariate Longitudinal Data}
\usage{
mnt_long_1d(
  yyij_new,
  ttij_new,
  nobs_new,
  pattern,
  side = "upward",
  chart = "CUSUM",
  method = "standard",
  parameter = 0.5,
  control_limit = Inf
)
}
\arguments{
\item{yyij_new}{a matrix of longitudinal observations. \cr
\code{yyij_new[i,j]} is the jth observation of the ith subject.}

\item{ttij_new}{a matrix of observation times. \cr
\code{ttij_new[i,j]} is the jth observation time of the ith subject. \cr
\code{yyij_new[i,j]} is observed at \code{ttij[i,j]}.}

\item{nobs_new}{an integer vector for number of observations. \cr
\code{nobs_new[i]} is the number of observations for the ith subject.}

\item{pattern}{the estimated regular longitudinal pattern}

\item{side}{a string \cr
If \code{side="upward"}, control charts aim to detect upward shifts.\cr
If \code{side="downward"}, control charts aim to detect downward shifts.\cr
If \code{side="both"}, control charts aim to detect shifts in both sides.}

\item{chart}{a string \cr
If \code{chart="CUSUM"}, apply CUSUM charts
If \code{chart="EWMA"}, apply EWMA charts}

\item{method}{a string \cr
If \code{method="standard"},
If \code{method="decorrelation"},
If \code{method="sprint"},
If \code{method="distribution and standard"},
If \code{method="distribution and decorrelation"},
If \code{method="distribution and sprint"},
If \code{method="nonparametric and standard"},
If \code{method="nonparametric and decorrelation"},}

\item{parameter}{a numeric value. \cr
\code{parameter} is the allowance constant if \code{method} is a CUSUM chart. \cr
\code{parameter} is the weighting parameter if \code{method} is an EWMA chart.}

\item{control_limit}{a numeric value
The control limit}
}
\description{
Monitor Univariate Longitudinal Data
}
