% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnt_long_md.R
\name{mnt_long_md}
\alias{mnt_long_md}
\title{Monitor Multivariate Longitudinal Data}
\usage{
mnt_long_md(
  yyijk_new,
  ttij_new,
  nobs_new,
  pattern,
  side = "both",
  method = "multivariate EWMA",
  parameter = 0.5,
  control_limit = Inf
)
}
\arguments{
\item{yyijk_new}{an array of longitudinal observations. \cr
\code{yyijk_new[i,j,k]} is the jth observation of the kth dimension of the ith subject.}

\item{ttij_new}{a matrix of observation times. \cr
\code{ttij_new[i,j]} is the jth observation time of the ith subject. \cr
\code{yyijk_new[i,j,]} is observed at \code{ttij[i,j]}.}

\item{nobs_new}{an integer vector for number of observations. \cr
\code{nobs_new[i]} is the number of observations for the ith subject.}

\item{pattern}{the estimated regular longitudinal pattern}

\item{side}{a string \cr
If \code{side="upward"}, control charts aim to detect upward shifts.\cr
If \code{side="downward"}, control charts aim to detect downward shifts.\cr
If \code{side="both"}, control charts aim to detect shifts in both sides.}

\item{method}{a string \cr
If \code{method="simultaneous CUSUM"}
If \code{method="simultaneous EWMA"}
If \code{method="multivariate CUSUM"}
If \code{method="multivariate EWMA"}
If \code{method="decorrelation CUSUM"}
If \code{method="decorrelation EWMA"}
If \code{method="nonparametric CUSUM"}
If \code{method="nonparametric EWMA"}}

\item{parameter}{a numeric value. \cr
\code{parameter} is the allowance constant if \code{method} is a CUSUM chart. \cr
\code{parameter} is the weighting parameter if \code{method} is an EWMA chart.}

\item{control_limit}{a numeric value. \cr
The control limit}
}
\description{
Monitor Multivariate Longitudinal Data
}
