% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnt_long_surv.R
\name{mnt_long_surv}
\alias{mnt_long_surv}
\title{Monitor Longitudinal Data for Survival Outcomes}
\usage{
mnt_long_surv(
  yyijk_new,
  ttij_new,
  nobs_new,
  pattern,
  method,
  parameter = 0.2,
  control_limit = Inf
)
}
\arguments{
\item{yyijk_new}{an array of longitudinal observations. \cr
\code{yyijk_new[i,j,k]} is the jth observation of the kth dimension of the ith subject.}

\item{ttij_new}{a matrix of observation times. \cr
\code{ttij_new[i,j]} is the jth observation time of the ith subject. \cr
\code{yyijk_new[i,j,]} is observed at \code{ttij[i,j]}.}

\item{nobs_new}{an integer vector for number of observations. \cr
\code{nobs_new[i]} is the number of observations for the ith subject.}

\item{pattern}{the estimated longitudinal and survival pattern}

\item{method}{a string \cr
If \code{method="risk"}
If \code{method="joint"}}

\item{parameter}{a numeric value. \cr
The weighting parameter in the modified EWMA charts.}

\item{control_limit}{a numeric value. \cr
The control limit}
}
\description{
Monitor Longitudinal Data for Survival Outcomes
}
