% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_search_control_limit.R
\name{search_control_limit}
\alias{search_control_limit}
\title{Search Control Limit}
\usage{
search_control_limit(
  chartij,
  ttij,
  nobs,
  starttime,
  endtime,
  ttmin,
  ttmax,
  ntimepoints,
  ATS_nominal,
  limit_lower,
  limit_step,
  limit_upper,
  no_signal_action = "omit",
  ATS_tol,
  limit_tol
)
}
\arguments{
\item{chartij}{a matrix of charting statistics. \cr
\code{chartij[i,j]} is the jth charting statistic of the ith subject.}

\item{ttij}{a matrix of observation times.  \cr
\code{ttij[i,j]} is the jth observation time of the ith subject.}

\item{nobs}{an integer vector for number of observations. \cr
\code{nobs[i]} is the number of observations for the ith subject.}

\item{starttime}{a vector of times from the start of monitoring. \cr
\code{starttime[i]} is the time that the ith subject starts to be monitored.}

\item{endtime}{a vector of times from the start of monitoring. \cr
\code{endtime[i]} is the time that the ith subject is lost to be monitored.}

\item{ttmin, ttmax}{two numeric values. \cr
[\code{ttmin},\code{ttmax}] is the design time interval.}

\item{ntimepoints}{an integer value. \cr
\code{ntimepoints} is the number of basic timepoints in the design time interval.}

\item{ATS_nominal}{a numeric value. \cr
\code{ATS_nominal} is the targeted/nominal ATS one wished to achieve.}

\item{ATS_tol}{a numeric value. \cr
Error tolerence for ATS.}

\item{limit_tol}{a numeric value. \cr
Error tolerence for control limit.}

\item{cl_lower, cl_step, cl_upper}{three numeric values. \cr
The control limit will be searched within the interval [\code{cl_lower},\code{cl_upper}]. \cr
When applying grid search, the algorithm will use a step size of \code{limit_step}. \cr
(Namely, the algorithm will start with \code{cl_lower}, 
and search through the sequences \code{cl_lower}, \code{cl_lower+cl_step}, \code{cl_lower+2*cl_step}, ... until \code{cl_upper}.)}
}
\value{
control limit, a numeric value.
}
\description{
Search Control Limit
}
\author{
LY
}
